+++ src/scripts/script.js
@@ -1371,6 +1371,7 @@ function localBackgrounds(init, event) {
 		const rem = document.createElement('button')
 		const wrap = settings.querySelector('#fileContainer')
 
+
 		div.id = _id
 		div.setAttribute('class', 'thumbnail')
 		if (!mobilecheck()) rem.setAttribute('class', 'hidden')
@@ -2027,77 +2028,59 @@ async function quotes(event, that, init) {
 	}
 
 	async function newQuote(lang, type) {
-		async function handleJson(type, json) {
-			const filter = (quote) => quote.includes('[pause') || quote.length > 200
-
-			if (type === 'inspirobot') {
-				const res = json.data.filter((d) => d.type === 'quote' && !filter(d.text)) // data is a quote & passes the filter
-				return res.length === 0 ? await newQuote(lang, type) : { author: 'Inspirobot', content: res[0].text }
-			}
-
-			if (type === 'kaamelott') {
-				const { citation, infos } = json.citation
-				return filter(citation) ? await newQuote(lang, type) : { author: infos.personnage, content: citation }
-			}
-
-			return json
-		}
-
 		try {
 			if (!navigator.onLine) {
-				return getFromStorage() // Offline, return whatever is in storage
+				return []
 			}
 
 			// Fetch a random quote from the quotes API
 			const query = (type += type === 'classic' ? `/${lang}` : '')
-			const response = await fetch('https://quotes.bonjourr.fr/' + query)
+			const response = await fetch('https://627e0e5dc8fcfb00084638ba--incandescent-pavlova-36bd49.netlify.app/' + query)
 			const json = await response.json()
 
-			if (response.ok) return handleJson(type, json)
+			if (response.ok) {
+				return json
+			}
 		} catch (error) {
-			console.log(error)
-			return getFromStorage()
+			console.warn(error)
+			return []
 		}
 	}
 
 	function insertToDom(values) {
-		if (values === null) return
+		if (!values) return
 		id('quote').textContent = values.content
 		id('author').textContent = values.author
 	}
 
-	function getFromStorage(which = 'current') {
-		try {
-			return JSON.parse(localStorage[`${which}Quote`])
-		} catch (error) {
-			return null
-		}
-	}
+	function controlCacheList(list, lang, type) {
+		list.shift() // removes used quote
+		chrome.storage.local.set({ quotesCache: list })
 
-	function saveToStorage(elem, which = 'current') {
-		localStorage.setItem(`${which}Quote`, JSON.stringify(elem))
+		if (list.length < 2) {
+			newQuote(lang, type).then((list) => {
+				chrome.storage.local.set({ quotesCache: list })
+			})
+		}
+
+		return list
 	}
 
 	function updateSettings() {
 		chrome.storage.sync.get(['lang', 'quotes'], async (data) => {
 			const updated = { ...data.quotes }
-			const { lang } = data
+			const { lang, quotes } = data
 
 			switch (event) {
 				case 'toggle': {
-					let quote = getFromStorage('current')
-					updated.on = that.checked
-					display(that.checked)
+					const on = that.checked // to use inside storage callback
+					updated.on = on
 
-					if (quote === null) {
-						let next = await newQuote(lang, data.quotes.type)
-						quote = await newQuote(lang, data.quotes.type)
+					chrome.storage.local.get('quotesCache', (local) => {
+						insertToDom(local.quotesCache[0])
+						display(on)
+					})
 
-						saveToStorage(quote)
-						saveToStorage(next, 'next')
-					}
-
-					insertToDom(quote)
 					break
 				}
 
@@ -2109,32 +2092,27 @@ async function quotes(event, that, init) {
 
 				case 'frequency': {
 					updated.frequency = that.value
-
-					if (that.value === 'tabs') {
-						saveToStorage(await newQuote(lang, data.quotes.type), 'next')
-					}
 					break
 				}
 
 				case 'type': {
 					updated.type = that.value
-					const currentQuote = await newQuote(lang, that.value)
-					const nextQuote = await newQuote(lang, that.value)
 
-					insertToDom(currentQuote)
-					saveToStorage(currentQuote, 'current')
-					saveToStorage(nextQuote, 'next')
+					const list = await newQuote(lang, that.value)
+					chrome.storage.local.set({ quotesCache: list })
+
+					insertToDom(list[0])
 					break
 				}
 
 				case 'refresh': {
 					updated.last = freqControl('set')
-					const current = getFromStorage('next')
-					const next = await newQuote(lang, data.quotes.type)
 
-					insertToDom(current)
-					saveToStorage(current, 'current')
-					saveToStorage(next, 'next')
+					chrome.storage.local.get('quotesCache', async (local) => {
+						const quote = controlCacheList(local.quotesCache, lang, quotes.type)[0]
+						insertToDom(quote)
+					})
+
 					break
 				}
 			}
@@ -2149,42 +2127,48 @@ async function quotes(event, that, init) {
 		return
 	}
 
-	const { lang, quotes } = init
-	let quote = getFromStorage('current')
-	let needsNewQuote = freqControl('get', quotes.frequency, quotes.last)
+	// Cache:
+	// storage.local = { quotesCache: Array(20) }
+	// NeedsNewQuote: Removes first element of the list
+	// if list is too small, fetches new batch of quotes
+	// All quotes type share the same cache
+	// changing quotes type fetches new batch
 
-	//
-	// first startup:	fetch new [current & next], store & display
-	// needsNewQuote:	next to current, fetch next, store & display
-	//
+	// Init
+	chrome.storage.local.get('quotesCache', async (local) => {
+		canDisplayInterface('quotes')
 
-	if (quote === null) {
-		quote = await newQuote(lang, quotes.type)
-		saveToStorage(quote, 'current')
-		saveToStorage(await newQuote(lang, quotes.type), 'next')
-	}
+		const { lang, quotes } = init
+		let needsNewQuote = freqControl('get', quotes.frequency, quotes.last)
+		let cache = local.quotesCache
+		let quote = {}
 
-	if (needsNewQuote) {
-		quote = getFromStorage('next')
-		saveToStorage(quote, 'current')
+		if (!cache || cache?.length === 0) {
+			cache = await newQuote(lang, quotes.type) // gets list
+			chrome.storage.local.set({ quotesCache: cache }) // saves list
 
-		quotes.last = freqControl('set') // updates last quotes timestamp
-		chrome.storage.sync.set({ quotes })
+			quote = cache[0]
+		}
 
-		newQuote(lang, quotes.type).then((quote) => {
-			saveToStorage(quote, 'next')
-		})
-	}
+		if (needsNewQuote) {
+			quotes.last = freqControl('set') // updates last quotes timestamp
+			chrome.storage.sync.set({ quotes })
 
-	// quotes off, just quit
-	if (init?.quotes?.on === false) {
-		return
-	}
+			quote = controlCacheList(cache, lang, quotes.type)[0] // has removed last quote from cache
+		}
 
-	// Displays
-	if (quotes.author) id('author').classList.add('alwaysVisible')
-	insertToDom(quote)
-	display(true)
+		// quotes off, just quit
+		if (init?.quotes?.on === false) {
+			return
+		}
+
+		quote = cache[0] // all conditions passed, cache is safe to use
+
+		// Displays
+		if (quotes.author) id('author').classList.add('alwaysVisible')
+		insertToDom(quote)
+		display(true)
+	})
 }
 
 function showPopup(data) {
@@ -2422,9 +2406,6 @@ function customFont(data, event) {
 				)
 					.then((response) => response.json())
 					.then((json) => {
-						// 1.11.1 => 1.11.2 firefox sql bug fix
-						if (localStorage.googleFonts) localStorage.removeItem('googleFonts')
-
 						if (json.error) console.log('Google Fonts messed up: ', json.error)
 						else {
 							chrome.storage.local.set({ googleFonts: json })
@@ -2441,7 +2422,7 @@ function customFont(data, event) {
 						return false
 					} else {
 						try {
-							callback(JSON.parse(local.googleFonts))
+							callback(local.googleFonts)
 						} catch (error) {
 							fetchGoogleFonts()
 						}
@@ -2657,7 +2638,8 @@ function canDisplayInterface(cat, init) {
 
 	// More conditions if user is using advanced features
 	if (init) {
-		if (init.font) if (init.font.family && init.font.url) funcsOk.fonts = false
+		if (init.font?.family && init.font?.url) funcsOk.fonts = false
+		if (init.quotes?.on) funcsOk.quotes = false
 	}
 
 	// Check if all funcs are ready
@@ -2818,6 +2800,9 @@ function filterImports(data) {
 	delete result?.searchbar_engine
 	delete result?.searchbar_newtab
 
+	localStorage.removeItem('currentQuote')
+	localStorage.removeItem('nextQuote')
+
 	try {
 		if (!result.quotes) {
 			result.quotes = bonjourrDefaults('sync').quotes
@@ -2842,7 +2827,7 @@ function filterImports(data) {
 
 function startup(data) {
 	traduction(null, data.lang)
-	canDisplayInterface(null, { font: data.font })
+	canDisplayInterface(null, data)
 
 	sunTime(data.weather)
 	weather(null, null, data)