+++ src/scripts/settings.js
@@ -148,7 +148,7 @@ function initParams(data, settingsDom) {
 	//
 
 	const bgfile = paramId('i_bgfile')
-	const uploadContainer = paramId('uploadContainer')
+	const fileContainer = paramId('fileContainer')
 
 	enterBlurs(paramId('i_favicon'))
 	enterBlurs(paramId('i_tabtitle'))
@@ -156,15 +156,15 @@ function initParams(data, settingsDom) {
 
 	// file input animation
 	bgfile.addEventListener('dragenter', function () {
-		uploadContainer.classList.add('dragover')
+		fileContainer.classList.add('dragover')
 	})
 
 	bgfile.addEventListener('dragleave', function () {
-		uploadContainer.classList.remove('dragover')
+		fileContainer.classList.remove('dragover')
 	})
 
 	bgfile.addEventListener('drop', function () {
-		uploadContainer.classList.remove('dragover')
+		fileContainer.classList.remove('dragover')
 	})
 
 	//general